+++ src/main/java/blackjack/domain/Deck.java
@@ -1,5 +1,6 @@
 package blackjack.domain;
 
+import blackjack.domain.card.Card;
 import java.util.Queue;
 
 public class Deck {
