+++ src/main/java/blackjack/domain/ShuffledDeckFactory.java
@@ -1,5 +1,8 @@
 package blackjack.domain;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Shape;
+import blackjack.domain.card.Symbol;
 import java.util.ArrayDeque;
 import java.util.Arrays;
 import java.util.Collections;
