+++ src/main/java/blackjack/domain/card/Shape.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 /**
  * 카드의 모양에 대한 정보를 담당하는 책임을 가지고 있습니다
