+++ src/main/java/blackjack/domain/card/Symbol.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 /**
  * 카드의 점수에 대한 정보를 담고 있는 클래스 입니다
