+++ src/main/java/blackjack/domain/participant/Dealer.java
@@ -1,8 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 public class Dealer extends Participant {
 
-    private static final int BUST_POINT = 21;
     private static final int CARD_DRAW_POINT = 16;
