+++ src/main/java/blackjack/response/PlayerCardsResponse.java
@@ -1,6 +1,6 @@
 package blackjack.response;
 
-import blackjack.domain.Player;
+import blackjack.domain.participant.Player;
 import java.util.List;
 import java.util.stream.Collectors;
