+++ src/main/java/blackjack/response/PlayersCardsResponse.java
@@ -1,7 +1,7 @@
 package blackjack.response;
 
-import blackjack.domain.Player;
-import blackjack.domain.Players;
+import blackjack.domain.participant.Player;
+import blackjack.domain.participant.Players;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
